Updated / Removed old listeners
LLPlayerListener and TMPlayerListener removed as they're obsolete. CompositeSummaryLogger no longer needed.

GameListener now accounts for null returns (for metrics that don't need to report anything per turn but need to compute something every turn for the final GAME_OVER report).

Improved output for TAGOccurrenceStatSummart (sorted by value). LoveLetter winnnigCard metric moved to metrics.

TODO:
 + PandemicGame -> uses enums. Needs to load from file.
 + Advantage003 use of metrics -> to be implemented as PlayerListener.
