+++ data/metrics/loveletter.json
@@ -5,12 +5,13 @@
       "class" : "evaluation.loggers.SummaryLogger"
     },
     [
-      {"class": "evaluation.metrics.MiscGameMetrics$ActionSimple" },
+      {"class": "evaluation.metrics.MiscGameMetrics$ActionSample" },
       {"class": "evaluation.metrics.GameMetrics$PlayerType" },
       {"class": "evaluation.metrics.GameMetrics$OrdinalPosition" },
       {"class": "evaluation.metrics.GameMetrics$GameScore" },
       {"class": "evaluation.metrics.GameMetrics$ActionSpace" },
-      {"class": "games.loveletter.stats.LoveLetterMetrics$ActionsPlayedWin" }
+      {"class": "games.loveletter.stats.LoveLetterMetrics$ActionsPlayedWin" },
+      {"class": "games.loveletter.stats.LoveLetterMetrics$WinningCards" }
     ]
   ]
 }
