--- data/terraformingmars/listeners/playerlistener.json
@@ -1,13 +1,0 @@
-{
-  "class" : "games.terraformingmars.stats.TMPlayerListener",
-  "args" : [
-    [
-      { "class" : "utilities.SummaryLogger" },
-      { "class" : "utilities.SummaryLogger" },
-      { "class" : "utilities.SummaryLogger" },
-      { "class" : "utilities.SummaryLogger" },
-      { "class" : "utilities.SummaryLogger" }
-    ],
-    { "class" : "utilities.SummaryLogger" }
-  ]
-}
