+++ src/main/java/core/AbstractGameState.java
@@ -201,7 +201,7 @@ public abstract class AbstractGameState {
             // If there is any information only available in History that could legitimately be used, then this should
             // be incorporated in the game-specific data in GameState where the correct hiding protocols can be enforced.
         }
-        // TODO: uncomment
+
         s.actionsInProgress = new Stack<>();
         actionsInProgress.forEach(
                 a -> s.actionsInProgress.push(a.copy())
